Валидация входящих параметров скрипта вынесена в validation.py
